fix load fvecs repository

+++ python/src/deglib/repository.py
@@ -26,4 +26,4 @@ def fvecs_read(filename: str | Path) -> np.ndarray:
     Taken from https://github.com/facebookresearch/faiss/blob/main/benchs/datasets.py#L12
     The loaded dataset should be in the format described here: http://corpus-texmex.irisa.fr/
     """
-    return ivecs_read(filename).view(np.int32)
+    return ivecs_read(filename).view(np.float32)
